requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
firebase-admin>=6.5.0
python-dotenv>=1.0
pydantic>=2.7
